test: login 함수 helpers 디렉토리로 분리
